+++ src/main/java/com/example/world2/model/entities/earth/liveMatter/animal/mammal/fourLegsOwner/rodent/Mouse.java
@@ -11,7 +11,7 @@ public class Mouse extends Rodent implements Climbable {
 
     public Mouse(String name, double maxLifetimeInSeconds, GeolocationPosition currentPosition, Earth earth, double weightInGrams,
                  double ageInSeconds, double widthInMillimeters, double heightInMillimeters, double jumpToBodySizeRatio,
-                 boolean warmBlooded, FoodType foodType, SkinType skinType, double movement, boolean isMakingNoise) {
+                 boolean warmBlooded, FoodType foodType, SkinType skinType, int movement, boolean isMakingNoise) {
         super(name, maxLifetimeInSeconds, currentPosition, earth, weightInGrams, ageInSeconds, widthInMillimeters,
                 heightInMillimeters, jumpToBodySizeRatio, warmBlooded, foodType, skinType, movement, isMakingNoise);
     }
